refact: changing adapter to androidx adapter with diffutil, extracted it and refactored agenda presentation layer
