--- agenda/src/main/java/com/picpay/agenda/presentation/home/UserListDiffCallback.kt
@@ -1,32 +1,0 @@
-package com.picpay.agenda.presentation.home
-
-import androidx.recyclerview.widget.DiffUtil
-import com.picpay.domain.entity.User
-
-class UserListDiffCallback(
-    private val oldList: List<User>,
-    private val newList: List<User>
-) : DiffUtil.Callback() {
-
-    override fun areItemsTheSame(
-        oldItemPosition: Int,
-        newItemPosition: Int
-    ): Boolean {
-        return oldList[oldItemPosition].username.equals(newList[newItemPosition].username)
-    }
-
-    override fun getOldListSize(): Int {
-        return oldList.size
-    }
-
-    override fun getNewListSize(): Int {
-        return newList.size
-    }
-
-    override fun areContentsTheSame(
-        oldItemPosition: Int,
-        newItemPosition: Int
-    ): Boolean {
-        return true
-    }
-}
