+++ app/src/test/java/com/picpay/desafio/android/presentation/di/AgendaModuleTest.kt
@@ -2,7 +2,7 @@ package com.picpay.desafio.android.presentation.di
 
 import android.app.Application
 import com.picpay.agenda.presentation.di.agendaModule
-import com.picpay.agenda.presentation.home.AgendaViewModel
+import com.picpay.agenda.presentation.home.contacts.AgendaViewModel
 import com.picpay.data.di.databaseModule
 import com.picpay.data.di.networkModule
 import com.picpay.desafio.android.BuildConfig
